b2b-claim-portal-api-development/requirements.txt
Flask-Cors==3.0.7
passlib==1.7.1
Flask-RESTful==0.3.7
Flask-JWT-Extended==3.18.1
ansimarkup==1.4.0
astroid==2.1.0
Autologging==1.3.1
autopep8==1.4.3
better-exceptions-fork==0.2.1.post6
Cerberus==1.2
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
Click==7.0
client==0.0.1
colorama==0.4.1
dependency-injector==3.14.5
Flask==1.0.2
Flask-HTTPAuth==3.3.0
Flask-Injector==0.11.0
Flask-Testing==0.7.1
Flask-WTF==0.14.2
gevent==1.4.0
greenlet==0.4.15
httplib2==0.12.1
idna==2.8
injector==0.12.0
isort==4.3.8
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
lxml==4.3.1
magicmock==0.3
MarkupSafe==1.1.1
mccabe==0.6.1
mock==2.0.0
parse==1.11.1
passlib==1.7.1
pbr==5.1.3
pep8==1.7.1
pycodestyle==2.5.0
pycparser==2.19
pycryptodome==3.8.1
Pygments==2.3.1
pylint==2.2.2
pytz==2018.9
requests==2.21.0
rpclib==2.8.0rc0
six==1.12.0
soaplib==1.0.0
spyne==2.12.16
typing==3.6.6
urllib3==1.25.*
Werkzeug==0.15.3
wrapt==1.11.1
WTForms==2.2.1
Flask-RESTful==0.3.7
numpy==1.16.3		
pandas==0.24.2
sendgrid==6.0.4	
twilio==6.26.1
ddtrace==0.27.1
